re_selection_panel
Module visible_time_range_ui

Functions

current_range_ui 🔒

query_range_ui 🔒

Draws ui for showing and configuring a query range.

resolved_visible_history_boundary_ui 🔒

save_visible_time_ranges 🔒

show_visual_time_range 🔒

time_range_editor 🔒

visible_history_boundary_combo_label 🔒

visible_history_boundary_ui 🔒

visible_time_range_ui 🔒

Draws ui for a visible time range from a given override path and a resulting query range.

visible_time_range_ui_for_data_result

visible_time_range_ui_for_view